+++ backend/garpix_user/views/logout_view.py
@@ -1,3 +1,6 @@
+from django.contrib.auth import get_user_model
+from garpix_utils.logs.enums.get_enums import Action, ActionResult
+from garpix_utils.logs.services.logger_iso import LoggerIso
 from rest_framework import parsers, renderers
 from garpix_user.models.access_token import AccessToken as Token
 from oauth2_provider.models import AccessToken
@@ -24,6 +27,17 @@ class LogoutView(APIView):
                 return Response({
                     'result': True,
                 })
+
+            message = f'Пользователь {request.user.username} вышел из системы.'
+            log = LoggerIso.create_log(action=Action.user_logout.value,
+                                       obj=get_user_model().__name__,
+                                       obj_address=request.path,
+                                       result=ActionResult.success,
+                                       sbj=request.user.username,
+                                       sbj_address=LoggerIso.get_client_ip(request),
+                                       msg=message)
+
+            LoggerIso.write_string(log)
         return Response({
             'result': False,
         }, status=401)